+++ settings.ini
@@ -1,6 +1,7 @@
 # パス設定
 [PATH]
-root_path = C:\Users\hukuyori\Desktop\CM_Analysis
+root_path = C:\Users\hukuyori\CM_Analysis
 video_path = Data\Movie
 cmData_path = Data\csv_file\MovieList_200508.csv
 result_path = Result\Cut
+ansData_path = Data\csv_file\Answer_cut_point.csv
